refactor: move data preview and table insertion click handlers from NotebookInteractions to AnalysisShell
